client/src/app/pages/join-match/join-match.component.spec.ts
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { Component, Input } from '@angular/core';
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { FormsModule } from '@angular/forms';
import { Router } from '@angular/router';
import { LogoComponent } from '@app/components/logo/logo.component';
import { MAX_ACCESS_CODE_LENGTH } from '@app/constants/constants';
import { AppMaterialModule } from '@app/modules/material.module';
import { JoinMatchService } from '@app/services/join-match/join-match.service';
import { MatchPlayerService } from '@app/services/match-player-service/match-player.service';
import { of } from 'rxjs';
import { JoinMatchComponent } from './join-match.component';

describe('JoinMatchComponent', () => {
    let component: JoinMatchComponent;
    let fixture: ComponentFixture<JoinMatchComponent>;
    let joinMatchService: jasmine.SpyObj<JoinMatchService>;
    let matchPlayerService: jasmine.SpyObj<MatchPlayerService>;
    let router: jasmine.SpyObj<Router>;
    @Component({
        selector: 'app-login',
    })
    class MockLoginComponent {
        @Input() title: string;
        @Input() text: string;
        @Input() inputType: string;
        @Input() label: string;
    }

    beforeEach(() => {
        joinMatchService = jasmine.createSpyObj('JoinMatchService', ['isValidAccessCode', 'isMatchAccessible', 'validatePlayerName', 'removeSpace']);
        matchPlayerService = jasmine.createSpyObj('MatchPlayerService', ['hasJoinMatch']);
        router = jasmine.createSpyObj('Router', ['navigateByUrl']);

        TestBed.configureTestingModule({
            declarations: [JoinMatchComponent, MockLoginComponent, LogoComponent],
            imports: [HttpClientTestingModule, AppMaterialModule, FormsModule],
            providers: [
                { provide: JoinMatchService, useValue: joinMatchService },
                { provide: MatchPlayerService, useValue: matchPlayerService },
                { provide: Router, useValue: router },
            ],
        });

        fixture = TestBed.createComponent(JoinMatchComponent);
        component = fixture.componentInstance;
        spyOn(window, 'alert').and.stub();
    });

    it('should create', () => {
        expect(component).toBeTruthy();
    });

    it('ngOnInit should set maxLength', () => {
        fixture.detectChanges();
        expect(joinMatchService.maxLength).toEqual(MAX_ACCESS_CODE_LENGTH);
    });

    it('should call onAccessCodeEntry when Enter key is pressed', () => {
        const event = new KeyboardEvent('keydown', { key: 'Enter' });
        spyOn(component, 'onAccessCodeEntry').and.callFake(() => {
            return;
        });

        window.dispatchEvent(event);

        expect(component.onAccessCodeEntry).toHaveBeenCalled();
    });

    it('onEnterKey should call onJoinMatch when Enter key is pressed and accessCodeIsValid is true', () => {
        const event = new KeyboardEvent('keydown', { key: 'Enter' });
        component.accessCodeIsValid = true;
        spyOn(component, 'onJoinMatch').and.callFake(() => {
            return;
        });
        window.dispatchEvent(event);

        expect(component.onJoinMatch).toHaveBeenCalled();
    });

    it('onAccessCodeEntry should set accessCodeIsValid to true if access code is valid', () => {
        joinMatchService.isValidAccessCode.and.returnValue(of(true));
        joinMatchService.isMatchAccessible.and.returnValue(of(true));

        component.onAccessCodeEntry();

        expect(component.accessCodeIsValid).toBe(true);
    });
    it('onAccessCodeEntry should display an error if the match is not accessible', () => {
        joinMatchService.isValidAccessCode.and.returnValue(of(true));
        joinMatchService.isMatchAccessible.and.returnValue(of(false));
        component.onAccessCodeEntry();

        expect(component.accessCodeIsValid).toBe(false);
        expect(window.alert).toHaveBeenCalled();
    });

    it('onAccessCodeEntry should set accessCodeError to true if access code is invalid', () => {
        joinMatchService.isValidAccessCode.and.returnValue(of(false));

        component.onAccessCodeEntry();

        expect(component.accessCodeError).toBe(true);
    });

    it('onJoinMatch should navigate to the correct URL when match is accessible and player name is valid', () => {
        joinMatchService.isValidAccessCode.and.returnValue(of(true));
        joinMatchService.isMatchAccessible.and.returnValue(of(true));
        joinMatchService.isValidAccessCode.and.returnValue(of(true));
        joinMatchService.validatePlayerName.and.returnValue(of(true));
        joinMatchService.input = 'ValidName';
        joinMatchService.accessCode = '1234';

        component.onJoinMatch();

        expect(router.navigateByUrl).toHaveBeenCalledWith(`play/wait/${joinMatchService.accessCode}`);
    });

    it('onJoinMatch should set nameError to true if player name is invalid', () => {
        joinMatchService.isValidAccessCode.and.returnValue(of(true));
        joinMatchService.isMatchAccessible.and.returnValue(of(true));
        joinMatchService.isValidAccessCode.and.returnValue(of(true));
        joinMatchService.validatePlayerName.and.returnValue(of(false));
        joinMatchService.input = 'InvalidName';

        component.onJoinMatch();

        expect(component.nameError).toBe(true);
    });

    it('onJoinMatch should show an alert if the match is not accessible', () => {
        joinMatchService.isValidAccessCode.and.returnValue(of(true));
        joinMatchService.isMatchAccessible.and.returnValue(of(false));
        joinMatchService.isValidAccessCode.and.returnValue(of(true));
        component.onJoinMatch();

        expect(window.alert).toHaveBeenCalledWith("L'organisateur a verrouillé cette partie");
    });
    it('onJoinMatch should reset parameter service if match is not accessible', () => {
        joinMatchService.isValidAccessCode.and.returnValue(of(false));
        component.onJoinMatch();

        expect(joinMatchService.maxLength).toEqual(MAX_ACCESS_CODE_LENGTH);
        expect(joinMatchService.input).toEqual('');
        expect(joinMatchService.accessCode).toEqual('');
        expect(matchPlayerService.hasJoinMatch).toEqual(false);
    });
});
